refactored from component structure
