+++ src/App.js
@@ -1,25 +1,15 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "./logo.svg";
+import "./App.css";
 // depreciated stylings throwing warnings -> not required
 // import 'bootstrap/dist/css/bootstrap.min.css';
-import Profile from './components/UserProfile/index.js'
+import Profile from "./components/UserProfile/index.js";
+import Form from "./components/Form";
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <div id="form">
+        <Form />
+      </div>
       <div id="profile">
         <Profile />
       </div>
